refactor: split business logic
